nova_backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.6.2
board==1.0
cachelib==0.12.0
camelot-py==0.11.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
config==0.5.1
cryptography==41.0.2
darr==0.5.4
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
ghostscript==0.7
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.2
MouseInfo==0.1.3
move==0.1.3
msgspec==0.18.6
mss==9.0.1
numpy==1.25.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
Pillow==10.0.0
psycopg2-binary==2.9.6
PyAutoGUI==0.9.54
pycparser==2.21
pydub==0.25.1
pygame==2.5.1
PyGetWindow==0.0.9
PyJWT==2.8.0
pymongo==4.6.1
PyMsgBox==1.0.9
pypdf==3.14.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sound==0.1.0
soundfile==0.12.1
soupsieve==2.4.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
square==0.0.3
tabulate==0.9.0
tk==0.1.0
trio==0.22.2
trio-websocket==0.10.3
TwoCaptcha==0.0.1
typing_extensions==4.10.0
tzdata==2023.3
undetected-chromedriver==3.5.0
urllib3==2.0.4
websockets==11.0.3
Werkzeug==2.3.4
wikipedia==1.4.0
wsproto==1.2.0
drf_yasg==1.21.7
